+++ src/day06.rs
@@ -113,16 +113,16 @@ fn parse_input(test: bool) -> Result<Map, Box<dyn error::Error>> {
     );
 }
 
-pub fn part1(test: bool) -> Result<u32, Box<dyn error::Error>> {
+pub fn part1(test: bool) -> Result<u64, Box<dyn error::Error>> {
     let mut values = parse_input(test)?;
     let mut explored: Vec<(i32, i32)> = vec![values.guard.position];
     while let Some(pos) = values.update() {
         explored.push(pos);
     }
-    return Ok(explored.iter().unique().count() as u32);
+    return Ok(explored.iter().unique().count() as u64);
 }
 
-pub fn part2(test: bool) -> Result<u32, Box<dyn error::Error>> {
+pub fn part2(test: bool) -> Result<u64, Box<dyn error::Error>> {
     let mut values = parse_input(test)?;
     let copy = values.clone();
     let obstructable = Arc::new(Mutex::new(0));
